requirements.txt
assemblyai
openai
pyaudio
elevenlabs
requests
pocketsphinx
python-dotenv